2004-08-08  Willem Jan Palenstijn <wjpalenstijn@users.sourceforge.net> 	* world/actors/AnimationTracker.*: a class to handle animations 	* world/actors/ActorAnimProcess.*: use AnimationTracker internally 	* world/actors/PathfindingProcess.*: set ACT_PATHFINDING flag 	* gumps/BarkGump.*: don't save string unnecessarily 	* kernel/Kernel.cpp: fixed double-terminate bug

+++ kernel/Kernel.cpp
@@ -290,8 +290,10 @@ void Kernel::killProcesses(ObjId objid, uint16 processtype)
 		Process* p = *it;
 
 		if ((objid == 0 || objid == p->item_num) &&
-			(processtype == 6 || processtype == p->type))
+			(processtype == 6 || processtype == p->type) && !p->terminated)
+		{
 			p->terminate();
+		}
 	}
 }
